:truck: Move types under directories with their name
